The first workble version.

+++ trunk/data_mining/proj1/defs.h
@@ -1,7 +1,7 @@
 #ifndef _typedef_header_
 #define _typedef_header_
 /* data type for gene feature */
-typedef float gene_feature_t; 
+typedef double gene_feature_t; 
 
 /* data type for class */
 typedef enum {positive, negative} gene_class_t;
